Clean up code structure and eliminate stupid SFINAE
